+++ sdk/informatica/arm-informaticadatamanagement/samples/v2/javascript/operationsListSample.js
@@ -1,0 +1,48 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+const { InformaticaDataManagement } = require("@azure/arm-informaticadatamanagement");
+const { DefaultAzureCredential } = require("@azure/identity");
+
+/**
+ * This sample demonstrates how to list the operations for the provider
+ *
+ * @summary list the operations for the provider
+ * x-ms-original-file: 2025-11-27/Operations_List_MaximumSet_Gen.json
+ */
+async function operationsList() {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "00000000-0000-0000-0000-000000000000";
+  const client = new InformaticaDataManagement(credential, subscriptionId);
+  const resArray = new Array();
+  for await (const item of client.operations.list()) {
+    resArray.push(item);
+  }
+
+  console.log(resArray);
+}
+
+/**
+ * This sample demonstrates how to list the operations for the provider
+ *
+ * @summary list the operations for the provider
+ * x-ms-original-file: 2025-11-27/Operations_List_MinimumSet_Gen.json
+ */
+async function operationsListMin() {
+  const credential = new DefaultAzureCredential();
+  const subscriptionId = "00000000-0000-0000-0000-000000000000";
+  const client = new InformaticaDataManagement(credential, subscriptionId);
+  const resArray = new Array();
+  for await (const item of client.operations.list()) {
+    resArray.push(item);
+  }
+
+  console.log(resArray);
+}
+
+async function main() {
+  await operationsList();
+  await operationsListMin();
+}
+
+main().catch(console.error);