+++ sdk/informatica/arm-informaticadatamanagement/src/classic/index.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export type { OperationsOperations } from "./operations/index.js";
+export type { OrganizationsOperations } from "./organizations/index.js";
+export type { ServerlessRuntimesOperations } from "./serverlessRuntimes/index.js";